refactor: validate 規則層抽離檔案系統相依，改吃注入的 IconSource
